Finalized implementation of asynchronous accept for SSL on Windows.
I've removed the unnecessary bakery counter and wait4Update_ method as the worker thread locks the fd/callback pairs vector and does not keep a local cache as I originally intended to do. That reduces complexity quite siginificantly on the code side..

I might have to make the vector a list and the mutex on it recursive, though, to allow for a callback to unregister the descriptor.