Laying the ground work for HTTP connections
We now have a functioning connection handler for HTTP, and a data handler that's called when data is ready. The next step would be to reduce the number of instances of the data connection to one, and then have it handle data, parsing it into requests and handing those requests to the request handler.