Adding a very, very tiny web server: Mesothelae
Mesothelae is a sub-order of the Arachnida order, to which spiders belong, and regroups a number of very rare spider species.

The Mesothelae program is intended to be a test bed to communicate with the likes of Internet Explorer, Firefox, Opera, etc. etc. to see what they need from a response.. For example: Firefox needs a Content-Length header for a 302 response or it will wait for content ad infinitum, not knowing that there is none - and thus won't go looking for the location the 302 response pointed to.