Laying the ground works for garbage-collecting dead connections
Dead connections can now be identified by the fact that their (new) status
flag is no longer good__, and are disconnected from the ConnectionHandler
when an exception is thrown by a data handler. Now, all that's left to do is
to garbage-collect the instance in the connection handler, which is the owner
of the connection object itself.