Garbage-collecting dead connections should now work
I haven't double-checked the correctness of this code yet, but it should be OK: when the handler throws an error, the connection's error handler is called which, in case it's the one installed by the HTTP connection handler, removes the connection from the list of connections it owns and destroys the connection itself. This recursively removes the connection from the connection handler as well, which should be enough to clean everything up.
It's not the prettiest code I've ever written, though...